backend/pyproject.toml
[project]
name = "strava-stats"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openpyxl>=3.1.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "tzdata>=2025.3",
]
